% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact_strength_metrics.R
\name{AS}
\alias{AS}
\title{Interaction strength asymmetry (AS)}
\usage{
AS(x, ...)
}
\arguments{
\item{x}{an adjacency matrix, which have one row and one column for each species in the bipartite network. The elements of the matrix can be any number, but in case of binary networks they will be either 0 or 1. A matrix element of 1 (or greater) signals that the respective column species (plants) and row species (frugivores) interact in the network.}

\item{...}{arguments to be passed to internal functions}
}
\value{
numeric
}
\description{
This function calculates interaction strength asymmetry for each specie in the network.
}
\examples{
net1 <- matrix(c(0, 1, 1, 1, 0, 1, 1, 1, 0, 0, 0, 1, 1, 1, 0, 1), nrow = 4, ncol = 4, dimnames = list(c("B1", "B2", "B3", "B4"), c("P1", "P2", "P3", "P4")))
AS(net1)

}
