% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeModules_spat.R
\name{computeModules.spat}
\alias{computeModules.spat}
\title{computeModules for spatial data}
\usage{
computeModules.spat(
  rh,
  rl,
  web,
  method = "Beckett",
  deep = FALSE,
  deleteOriginalFiles = TRUE,
  steps = 1e+06,
  tolerance = 1e-10,
  experimental = FALSE,
  forceLPA = FALSE
)
}
\arguments{
\item{web}{}
}
\description{
This function takes a bipartite weighted graph and computes modules by
applying Newman's modularity measure in a bipartite weighted version to it.
metaComputeModules re-runs the algorithm several times, returning the most
modular result, to stabilise modularity computation.
}
