% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested_spat.R
\name{nested.spat}
\alias{nested.spat}
\title{nested_spat metrics}
\usage{
nested.spat(
  rh,
  rl,
  web,
  method = "weighted NODF",
  rescale = FALSE,
  normalised = TRUE
)
}
\arguments{
\item{web}{A matrix with elements of a set (e.g. plants) as rows, elements of a second set (e.g. pollinators) as columns and number of interactions as entries.}

\item{method}{One or more of the following: \option{discrepancy}, \option{discrepancy2}, \option{binmatnest}, \option{NODF}, \option{NODF2}, \option{C score}, \option{checker}, \option{weighted NODF}, \option{wine}, \option{ALL}. See details for details on each method.}

\item{rescale}{Should all measures be rescaled so that higher values mean higher nestedness? Defaults to FALSE, i.e. the standard interpretation of each measure is maintained.}

\item{normalised}{Logical, defaulting to TRUE. Should C-scores be normalised to a value between 0 and 1? See \code{\link[bipartite]{C.score}} for details.}
}
\description{
blabla
}
