% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested.spat.R
\name{nested_vec}
\alias{nested_vec}
\title{Several nestedness metrics for each raster cell}
\usage{
nested_vec(x, web, hlyr, method = "NODF", rescale = FALSE, normalised = TRUE)
}
\arguments{
\item{x}{Vector. Each pixel of a raster (stack) containing presence-
absence data (0 or 1) for the species of higher or lower trophic levels}

\item{web}{Matrix. A weighted bipartite network matrix, binary (o or 1) or
not, where the lower level species (e.g. plants) are rows and higher level (e.g.
frugivores or pollinators)species are columns. The layers (species) of each
raster must be sorted according to the bipartite network order!}

\item{hlyr}{Logical vector indicating if the species are from higher level}

\item{method}{One or more of the following: \option{discrepancy}, \option{discrepancy2}, \option{binmatnest}, \option{NODF}, \option{NODF2}, \option{C score}, \option{checker}, \option{weighted NODF}, \option{wine}, \option{ALL}. See details for details on each method.}

\item{rescale}{Should all measures be rescaled so that higher values mean higher nestedness? Defaults to FALSE, i.e. the standard interpretation of each measure is maintained.}

\item{normalised}{Logical, defaulting to TRUE. Should C-scores be normalised to a value between 0 and 1? See \code{\link[bipartite]{C.score}} for details.}
}
\value{
Vector
}
\description{
Calculates any of several nestedness metrics available for
bipartite networks, which can be weighted or not, for each raster cell. The
calculation will be made for each raster cell, that is, for each subnetwork
formed by the co-occurrence modeled for the pairs of recorded interacting species.
}
\author{
Neander Marcel Heming and Cynthia Valéria Oliveira
}
