% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networklevel.spat.R
\name{networklevel.spat}
\alias{networklevel.spat}
\title{Several network indices at the network level for raster data.}
\usage{
networklevel.spat(
  rh,
  rl,
  web,
  index = "connectance",
  level = "both",
  weighted = T,
  ISAmethod = "Bluethgen",
  SAmethod = "Bluethgen",
  extinctmethod = "r",
  nrep = 100,
  CCfun = "median",
  dist = "horn",
  normalise = T,
  empty.web = T,
  logbase = "e",
  intereven = "prod",
  H2_integer = T,
  fcweighted = T,
  fcdist = "euclidean",
  legacy = F
)
}
\arguments{
\item{rh}{SpatRaster. A raster (stack) containing presence-absence data (0 or 1)
for the higher level set of species.}

\item{rl}{A SpatRaster. A raster (stack) containing presence-absence data (0 or 1)
for the lower level set of species.}

\item{web}{Matrix. A weighted bipartite network matrix, binary (o or 1) or
not, where the lower level species (e.g. plants) are rows and higher level (e.g.
frugivores or pollinators)species are columns. The layers (species) of each
raster must be sorted according to the bipartite network order!}

\item{index}{Vector of indices to be calculated for each trophic level of the web;
            options are: 
    \itemize{
       		\item \option{degree}, 
		\item \option{ND} for normalised degrees, 
		%\item \option{dependence}, 
		\item \option{species strength} as sum of dependencies for each species, 
		\item \option{nestedrank} as rank in a nested matrix,
		\item \option{interaction push pull} for interaction push/pull (our version of dependence asymmetry: see details),  
	        	\item \option{PDI} for Paired Differences Index, 
	        	\item \option{resource range} for Poisot et al. (2012)'s index of unused resources,
	        	\item \option{species specificity} (or coefficient of variation of interactions),
        		\item \option{PSI} for pollination service index (or pollinator support index, depending on the trophic level), 
	        	\item \option{NS} for node specialisation index, 
		\item \option{betweenness} for betweenness, 
		\item \option{closeness} (both automatically also return their weighted counterparts proposed by Tore Opsahl in package \pkg{tnet}),
        		\item \option{Fisher} for Fisher's alpha index, 
		\item \option{diversity} for Shannon diversity of interactions of that species, 
		\item \option{effective partners} for the effective number of interacting partners,  
		\item\option{proportional generality} a quantitative version of normalised degree,
		\item\option{proportional similarity} specialisation measured as similarity between use and availability,
		\item \option{d} for Blüthgen's d', 
		\item \option{ALL} for all the aforementioned. 
	        	\item \option{ALLBUTD} (default) excludes only the dependence matrix and leads to the output being simplified to a matrix per trophic level.}
            }

\item{level}{For which level(s) should the indices be computed? Options are \option{both} (default), \option{lower} and \option{higher}. Output for one trophic level will be returned as a data.frame (unless dependencies are also reported), while for two levels a list of two matrices (higher and lower) will be returned.}

\item{weighted}{Logical}

\item{ISAmethod}{Method to use for calculating interaction strength (= dependence) asymmetry; original by \option{Bascompte} is yielding artefact results based only on the asymmetry of the web (as shown by example in Blüthgen et al. 2007 analytically in Blüthgen 2010) and should hence be avoided; \option{Bluethgen} (default) excludes singletons and corrects for low number of interactions (range -1 to 1).}

\item{SAmethod}{How to aggregate d'-based specialisation values: mean of log-transformed dependencies (\option{log}) or Blüthgen's marginal totals-weighted mean (default); see Blüthgen et al. (2007).}

\item{extinctmethod}{Specifies how species are removed from matrix: \option{random}, \option{degree} or \option{abundance} (partial matching). See \code{\link[bipartite]{second.extinct}} for details an option to predefine the sequence externally; idea from Memmott et al. (2004).}

\item{nrep}{Number of replicates for the extinction sequence analysis.}

\item{CCfun}{Method to use when calculating the clustering coefficient. Originally proposed as mean of cluster coefficients for each species. Defaults to \option{median}, because cluster coefficients are strongly skewed. }

\item{dist}{Distance metric to be used to calculate niche overlap. Any of \pkg{vegan}'s \code{vegdist}-metrics can be used; defaults to Horn's index, which is the recommendation of Krebs (1989). Binary percent niche overlap would be computed with \cr  \option{dist = "jaccard"}.}

\item{normalise}{Logical; shall the C-score and togetherness metrics be normalised to a range of 0 to 1? Defaults to \code{TRUE}.}

\item{empty.web}{Shall the empty columns and rows be deleted? Defaults to TRUE. option FALSE not yet fully tested}

\item{logbase}{numeric; gives the base to which the various diversity indices (partner diversity, effective partners) shallb be calculated, typcially exp(1) (default) or 2. Log2 is the proposal for generality and vulnerability by Bersier et al. (2002), while Shannon typically uses the natural logarithm, ln. The choice of the base will not affect the results qualitatively, only by a scaling factor.}

\item{intereven}{Shall all cells of the matrix be used to calculate the interaction evenness \cr
  (\option{intereven = "prod"})? Or, as given by Bersier et al. (2002) and Tylianakis et al. (2007), should only the realised links be used (\option{intereven = "sum"}; default)? Prod and sum refer to using the log of the \bold{product} of matrix dimensions (i.e. all cells) or the log of the \bold{sum} of non-zero cells (i.e. number of links)  as denominator in the evenness formula. See last paragraph of the details-section for views on these two options!}

\item{H2_integer}{Logical; indicates whether values in web are integers. Passed on to \code{\link[bipartite]{H2fun}}; see there for details.}

\item{fcweighted}{Logical; when computing ``functional complementarity'' sensu function \code{\link[bipartite]{fc}}, should the weights of the matrix be used. Defaults to TRUE, but original paper (Devoto et al. 2012) is based on FALSE.}

\item{fcdist}{Distance measure to be used to compute functional complementarity through \code{\link[bipartite]{fc}}; any measure accepted by \code{\link{dist}} is acceptable.}

\item{legacy}{Logical; should the old (pre-2.00) version of \code{networklevel} be used? To be backward compatible, the old \code{networklevel}-function is still available (\code{.networklevel}) and can be called by setting \option{legacy=TRUE}. This is only for the transition period until all papers in the making have been published (or binned). Index names and sometimes unclear focal level were downsides of the old implementation, which is now remedied. Thus, the use of \option{legacy=TRUE} and the direct call of \code{.networklevel} are strongly discouraged!}
}
\value{
Spatraster with the choose network level metric
}
\description{
Calculate various indices and values for bipartite networks on
raster data. You can see the spatial variation of important metrics such as
connectance, web asymmetry, network specialization (H2) and four options to
compute nestedness. View all available indexes in Details.
}
\details{
Note that if a network is very small, with few nodes or links, it may be
impossible to calculate the choose metric or calculated result may be
unreliable. As the calculation is made with the subnet of each pixel, even in
cases like this, it is possible to visualize the spatialized metric on a
macroecological scale. We suggest that users select one index at a time for
time-efficient spatial calculation, being the default the "connectance". Also
note that some indices may have a high processing time for the entire raster,
depending on their algorithms and the processing capacity and RAM of the
machine used. The current available indices for network level metrics are
listed bellow and users can find more information about them at
\link[bipartite]{networklevel}.

There are metrics computed for each level:
• ‘connectance’,
• ‘web asymmetry’,
• ‘links per species’,
• ‘number of compartments’,
• ‘compartment diversity’,
• ‘cluster coefficient’,
• ‘nestedness’
• ‘NODF’,
• ‘weighted nestedness’
• ‘weighted NODF’,
• ‘ISA’ (or alternatively ‘interaction strength asymmetry’ or ‘dependence
asymmetry’),
• ‘SA’ (or alternatively ‘specialisation asymmetry’),
• ‘linkage density’,
• ‘weighted connectance’,
• ‘Fisher alpha’,
• ‘interaction evenness’,
• ‘Alatalo interaction evenness’,
• ‘Shannon diversity’,
• ‘H2’ (network specialization);

and/or those metrics invoked through "grouplevel":
• ‘number of species’ in the respective trophic level,
• ‘mean number of links’,
• ‘mean number of shared partners’,
• ‘weighted cluster coefficient’,
• ‘degree distribution’,
• ‘togetherness’,
• ‘C score’,
• ‘V ratio’,
• ‘discrepancy’,
• ‘extinction slope’.
}
\examples{
\dontrun{
library(terra)
library(bipartite)
# load bipartite network and the raster stacks of higher level and lower level
species
bipnet <- read.csv(system.file("extdata", "bipnet.csv",
package="net.raster"), row.names=1, sep= ";" ) #change separator if necessary
rasth <- rast(system.file("extdata", "rasth.tif",
package="net.raster"))
rastl <- rast(system.file("extdata", "rastl.tif",
package="net.raster"))
# applying the function to compute connectance (default) for both levels
(default), that is, the entire network
conn <- networklevel.spat (rasth, rastl, bipnet)
plot(conn)
# calculating  H2 for he entire network
webh2 <- networklevel.spat (rasth, rastl, bipnet, index="H2")
plot(webh2)
# computing degree distribution for one group, the lower level
lowdd <- networklevel.spat (rasth, rastl, bipnet, index="degree distribution",
level="lower")
plot(lowdd)
}
}
\references{
Carsten F. Dormann
Almeida-Neto, M., Loyola, R.D., Ulrich, W., Guimaraes, P., Guimaraes, Jr.,
P.R. 2008. A consistent metric for nestedness analysis in ecological systems:
reconciling concept and measurement. Oikos 117, 1227–1239
Almeida-Neto, M. & Ulrich, W. (2011) A straightforward computational approach
for measuringnestedness using quantitative matrices. Environmental Modelling
& Software 26, 173–178
Bascompte, J., Jordano, P. and Olesen, J. M. 2006 Asymmetric coevolutionary
networks facilitate biodiversity maintenance. Science 312, 431–433
Bersier, L. F., Banasek-Richter, C. and Cattin, M. F. (2002) Quantitative
descriptors of food-web matrices. Ecology 83, 2394–2407
Blüthgen, N. (2010) Why network analysis is often disconnected from community
ecology: A critique and an ecologist’s guide. Basic and Applied Ecology 11,
185–195
Blüthgen, N., Menzel, F., Hovestadt, T., Fiala, B. and Blüthgen N. 2007
Specialization, constraints and conflicting interests in mutualistic networks.
Current Biology 17, 1–6
Burgos, E., H. Ceva, R.P.J. Perazzo, M. Devoto, D. Medan, M. Zimmermann, and
A. Maria Delbue (2007) Why nestedness in mutualistic networks? Journal of
Theoretical Biology 249, 307–313
Corso G, de Araújo AIL, de Almeida AM (2008) A new nestedness estimator in
community networks.
Devoto M., Bailey S., Craze P., and Memmott J. (2012) Understanding and
planning ecological restoration of plant-pollinator networks. Ecology Letters
15, 319–328. http://dx.doi.org/10.1111/j.1461-0248.2012.01740.x
Dormann, C.F., Fründ, J., Blüthgen, N., and Gruber, B. (2009) Indices, graphs
and null models: analysing bipartite ecological networks. The Open Ecology
Journal 2, 7–24.
Dunne, J. A., R. J. Williams, and N. D. Martinez. 2002 Food-web structure
and network theory: the role of connectance and size. Proceedings of the
National Academy of Science USA 99, 12917–12922
Galeano, J., Pastor, J.M. and Iriondo, J.M. (2008) Weighted-Interaction
Nestedness Estimator (WINE): A new estimator to calculate over frequency
matrices.
Gotelli, N. J., and G. R. Graves. 1996 Null Models in Ecology. Smithsonian
Institution Press, Washington D.C.
Krebs, C. J. 1989. Ecological Methodology. Harper Collins, New York.
}
\seealso{
\code{\link{prep.web}}, \code{\link{nested.spat}},
\code{\link{computeModules.spat}}
}
\author{
Neander Marcel Heming and Cynthia Valéria Oliveira
}
