% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_web.R
\name{prep.web}
\alias{prep.web}
\title{prep_web}
\usage{
prep.web(rh, rl, web)
}
\arguments{
\item{rh}{SpatRaster of higher level species in the network}

\item{rl}{SpatRaster of lower level species in the network}

\item{web}{A bipartite network, which can be weighted or not}
}
\value{
A vector (?) with the species names of higher and lower levels
}
\description{
Checks whether the species names of each Spatraster (higher and lower level)
match the species names in the network.
}
\author{
Neander Marcel Heming and Cynthia Valéria Oliveira
}
