% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieslevel.spat.R
\name{sl_vec}
\alias{sl_vec}
\title{Several network indices at the species level for each raster cell}
\usage{
sl_vec(
  x,
  web,
  hlyr,
  index = "closeness",
  level = "both",
  weighted = F,
  logbase = exp(1),
  low.abun = NULL,
  high.abun = NULL,
  PDI.normalise = TRUE,
  PSI.beta = c(1, 0),
  nested.method = "NODF",
  nested.normalised = TRUE,
  nested.weighted = TRUE,
  empty.web = TRUE
)
}
\arguments{
\item{x}{Vector. Each pixel of a raster (stack) containing presence-
absence data (0 or 1) for the species of higher or lower trophic levels}

\item{web}{Matrix. A weighted bipartite network matrix, binary (o or 1) or
not, where the lower level species (e.g. plants) are rows and higher level (e.g.
frugivores or pollinators)species are columns. The layers (species) of each
raster must be sorted according to the bipartite network order!}

\item{hlyr}{Logical vector indicating if the species are from higher level}

\item{index}{Vector of indices to be calculated for each trophic level of the web;
            options are: 
    \itemize{
       		\item \option{degree}, 
		\item \option{ND} for normalised degrees, 
		%\item \option{dependence}, 
		\item \option{species strength} as sum of dependencies for each species, 
		\item \option{nestedrank} as rank in a nested matrix,
		\item \option{interaction push pull} for interaction push/pull (our version of dependence asymmetry: see details),  
	        	\item \option{PDI} for Paired Differences Index, 
	        	\item \option{resource range} for Poisot et al. (2012)'s index of unused resources,
	        	\item \option{species specificity} (or coefficient of variation of interactions),
        		\item \option{PSI} for pollination service index (or pollinator support index, depending on the trophic level), 
	        	\item \option{NS} for node specialisation index, 
		\item \option{betweenness} for betweenness, 
		\item \option{closeness} (both automatically also return their weighted counterparts proposed by Tore Opsahl in package \pkg{tnet}),
        		\item \option{Fisher} for Fisher's alpha index, 
		\item \option{diversity} for Shannon diversity of interactions of that species, 
		\item \option{effective partners} for the effective number of interacting partners,  
		\item\option{proportional generality} a quantitative version of normalised degree,
		\item\option{proportional similarity} specialisation measured as similarity between use and availability,
		\item \option{d} for Blüthgen's d', 
		\item \option{ALL} for all the aforementioned. 
	        	\item \option{ALLBUTD} (default) excludes only the dependence matrix and leads to the output being simplified to a matrix per trophic level.}
            }

\item{level}{For which level(s) should the indices be computed? Options are \option{both} (default), \option{lower} and \option{higher}. Output for one trophic level will be returned as a data.frame (unless dependencies are also reported), while for two levels a list of two matrices (higher and lower) will be returned.}

\item{logbase}{numeric; gives the base to which the various diversity indices (partner diversity, effective partners) shallb be calculated, typcially exp(1) (default) or 2. Log2 is the proposal for generality and vulnerability by Bersier et al. (2002), while Shannon typically uses the natural logarithm, ln. The choice of the base will not affect the results qualitatively, only by a scaling factor.}

\item{low.abun}{Optional vector of independent abundances of the \emph{lower} trophic level to be passed on to \code{\link[bipartite]{dfun}} and used as `resource availability' in indices proportional similarity and proportional generality.}

\item{high.abun}{Optional vector of independent abundances of the \emph{higher} trophic level to be passed on to \code{\link[bipartite]{dfun}} and used as `resource availability' in indices proportional similarity and proportional generality.}

\item{PDI.normalise}{Logical; should the normalised PDI be computed? Defaults to TRUE. See function \code{\link[bipartite]{PDI}} for details and reasoning.}

\item{PSI.beta}{A length 2 vector of parameter determining the effect of floral constancy and pollen mixing on the proportion of same-species pollen delivered to a plant. Defaults to \code{c(1, 0)}. See section details for details.}

\item{nested.method}{One of \option{NODF}, \option{binmatnest}, \option{wine} or \option{sort}. See \code{\link[bipartite]{nestedrank}} for details.}

\item{nested.normalised}{Logical; defaulting to TRUE. Divides the nestedrank-1 by the number of species -1, thereby ranging it between 0 (most generalist) and 1 (most specialised). Species sequences from different networks are now comparable.}

\item{nested.weighted}{Logical; should the ``nestedrank'' be computed based on weighted network (default) or binary version (FALSE)?}

\item{empty.web}{Shall the empty columns and rows be deleted? Defaults to TRUE. option FALSE not yet fully tested}
}
\value{
Vector
}
\description{
Calculate various simple numbers and indices related to network
nodes (species) for each pixel. The calculation will be made for each raster
cell, that is, for each subnetwork formed by the co-occurrence modeled for
the pairs of recorded interacting species.
}
\references{
Carsten F. Dormann & Jochen Fründ ("bipartite" package)
Alarcon, R., Waser, N.M. and Ollerton, J. 2008. Year-to-year variation in the
topology of a plantpollinator interaction network. Oikos 117, 1796–1807
Barrat, A., M. Barthélemy, R. Pastor-Satorras, and A. Vespignani. 2004. The
architecture of complex weighted networks. Proceedings of the National Academy
of Sciences of the USA 101, 3747—3752. doi: 10.1073/pnas.0400087101.
Bascompte, J., Jordano, P. and Olesen, J. M. (2006) Asymmetric coevolutionary
networks facilitate biodiversity maintenance. Science 312, 431–433
Berlow, E. L., A. M. Neutel, J. E. Cohen, P. C. de Ruiter, B. Ebenman, M.
Emmerson, J. W. Fox,V. A. A. Jansen, J. I. Jones, G. D. Kokkoris, D. O.
Logofet, A. J. McKane, J. M. Montoya & O.
Petchey (2004) Interaction strengths in food webs: issues and opportunities.
Journal of Animal Ecology 73, 585-–598
Blüthgen, N., Menzel, F. and Blüthgen, N. (2006) Measuring specialization
in species interaction networks. BMC Ecology 6, 9
Dormann, C.F. (2011) How to be a specialist? Quantifying specialisation in
pollination networks.Network Biology 1, 1–20
Feinsinger, P., Spears, E.E. and Poole,R. W. (1981) A simple measure of
niche breadth. Ecology 62, 27–32.
Julliard, R., Clavel, J., Devictor, V., Jiguet, F. and Couvet, D. (2006)
Spatial segregation of specialists and generalists in bird communities.
Ecology Letters 9, 1237-–1244
Martín Gonzáles, A.M., Dalsgaard, B. and Olesen, J.M. (2010) Centrality
measures and the importance of generalist species in pollination networks.
Ecological Complexity, 7, 36–43
Opsahl, T. & Panzarasa, P. (2009). Clustering in weighted networks. Social
Networks, 31, 155–163
Poisot, T., Lepennetier, G., Martinez, E., Ramsayer, J., and Hochberg, M.E.
(2011a) Resource availability affects the structure of a natural bacteria-
bacteriophage community. Biology Letters 7,201–204
Poisot, T., Bever, J.D., Nemri, A., Thrall, P.H., and Hochberg, M.E. (2011b)
A conceptual framework for the evolution of ecological specialisation.
Ecology Letters 14, 841–851
Poisot, T., E. Canard, N. Mouquet, and M. E. Hochberg (2012) A comparative
study of ecological specialization estimators. Methods in Ecology and
Evolution 3, 537-–544. doi: 10.1111/j.2041-210X.2011.00174.x.
Vázquez, D. P., Melian, C. J., Williams, N. M., Blüthgen N., Krasnov B. R.
and Poulin, R. (2007) Species abundance and asymmetric interaction strength
in ecological networks. Oikos 116, 1120–1127
}
\author{
Neander Marcel Heming and Cynthia Valéria Oliveira
}
