% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specieslevel.spat.R
\name{specieslevel.spat}
\alias{specieslevel.spat}
\title{Several network indices at the species level for raster data}
\usage{
specieslevel.spat(
  rh,
  rl,
  web,
  index = "closeness",
  level = "both",
  weighted = FALSE,
  logbase = exp(1),
  low.abun = NULL,
  high.abun = NULL,
  PDI.normalise = TRUE,
  PSI.beta = c(1, 0),
  nested.method = "NODF",
  nested.normalised = TRUE,
  nested.weighted = TRUE,
  empty.web = TRUE
)
}
\arguments{
\item{rh}{SpatRaster. A raster (stack) containing presence-absence data (0 or 1)
for the higher level set of species.}

\item{rl}{A SpatRaster. A raster (stack) containing presence-absence data (0 or 1)
for the lower level set of species.}

\item{web}{Matrix. A weighted bipartite network matrix, binary (o or 1) or
not, where the lower level species (e.g. plants) are rows and higher level (e.g.
frugivores or pollinators)species are columns. The layers (species) of each
raster must be sorted according to the bipartite network order!}

\item{index}{Vector of indices to be calculated for each trophic level of the web;
            options are: 
    \itemize{
       		\item \option{degree}, 
		\item \option{ND} for normalised degrees, 
		%\item \option{dependence}, 
		\item \option{species strength} as sum of dependencies for each species, 
		\item \option{nestedrank} as rank in a nested matrix,
		\item \option{interaction push pull} for interaction push/pull (our version of dependence asymmetry: see details),  
	        	\item \option{PDI} for Paired Differences Index, 
	        	\item \option{resource range} for Poisot et al. (2012)'s index of unused resources,
	        	\item \option{species specificity} (or coefficient of variation of interactions),
        		\item \option{PSI} for pollination service index (or pollinator support index, depending on the trophic level), 
	        	\item \option{NS} for node specialisation index, 
		\item \option{betweenness} for betweenness, 
		\item \option{closeness} (both automatically also return their weighted counterparts proposed by Tore Opsahl in package \pkg{tnet}),
        		\item \option{Fisher} for Fisher's alpha index, 
		\item \option{diversity} for Shannon diversity of interactions of that species, 
		\item \option{effective partners} for the effective number of interacting partners,  
		\item\option{proportional generality} a quantitative version of normalised degree,
		\item\option{proportional similarity} specialisation measured as similarity between use and availability,
		\item \option{d} for Blüthgen's d', 
		\item \option{ALL} for all the aforementioned. 
	        	\item \option{ALLBUTD} (default) excludes only the dependence matrix and leads to the output being simplified to a matrix per trophic level.}
            }

\item{level}{For which level(s) should the indices be computed? Options are \option{both} (default), \option{lower} and \option{higher}. Output for one trophic level will be returned as a data.frame (unless dependencies are also reported), while for two levels a list of two matrices (higher and lower) will be returned.}

\item{weighted}{Logical}

\item{logbase}{numeric; gives the base to which the various diversity indices (partner diversity, effective partners) shallb be calculated, typcially exp(1) (default) or 2. Log2 is the proposal for generality and vulnerability by Bersier et al. (2002), while Shannon typically uses the natural logarithm, ln. The choice of the base will not affect the results qualitatively, only by a scaling factor.}

\item{low.abun}{Optional vector of independent abundances of the \emph{lower} trophic level to be passed on to \code{\link[bipartite]{dfun}} and used as `resource availability' in indices proportional similarity and proportional generality.}

\item{high.abun}{Optional vector of independent abundances of the \emph{higher} trophic level to be passed on to \code{\link[bipartite]{dfun}} and used as `resource availability' in indices proportional similarity and proportional generality.}

\item{PDI.normalise}{Logical; should the normalised PDI be computed? Defaults to TRUE. See function \code{\link[bipartite]{PDI}} for details and reasoning.}

\item{PSI.beta}{A length 2 vector of parameter determining the effect of floral constancy and pollen mixing on the proportion of same-species pollen delivered to a plant. Defaults to \code{c(1, 0)}. See section details for details.}

\item{nested.method}{One of \option{NODF}, \option{binmatnest}, \option{wine} or \option{sort}. See \code{\link[bipartite]{nestedrank}} for details.}

\item{nested.normalised}{Logical; defaulting to TRUE. Divides the nestedrank-1 by the number of species -1, thereby ranging it between 0 (most generalist) and 1 (most specialised). Species sequences from different networks are now comparable.}

\item{nested.weighted}{Logical; should the ``nestedrank'' be computed based on weighted network (default) or binary version (FALSE)?}

\item{empty.web}{Shall the empty columns and rows be deleted? Defaults to TRUE. option FALSE not yet fully tested}
}
\value{
Spatraster object with the choose species level metric for both
trophic levels (default), higher or lower level.
}
\description{
Calculate various simple numbers and indices related to network
nodes (species) for raster data. You can see the spatial variation of important
metrics such as species strength and centralities (closeness - default metric,
betweenness, and degree). View all available indexes in Details and note that
users may select one index at a time for time-efficient spatial calculation.
For levels, "both" is the default, but users can also choose one level of
interest at a time (higher or lower), as well as choose unweighted metrics
whenever their calculation allows, with the default being weighted calculation.
}
\details{
Note that if a network is very small, with few nodes or links, it may be
impossible to calculate the choose metric or calculated result may be
unreliable. As the calculation is made with the subnet of each pixel, even in
cases like this, it is possible to visualize the spatialized metric on a
macroecological scale. Users may select one index at a time for time-efficient
spatial calculation,there will be an error message otherwise. Also note that
some indices may have a high processing time for the entire raster, depending
on their algorithms and the processing capacity and RAM of the machine used.
The current available indices for species level metrics are listed bellow and
users can find more information about them at  \link[bipartite]{specieslevel}
• ‘degree’,
• 'normalised degree',
• ‘species strength’,
• ‘nestedrank’,
• ‘interaction push pull’ - interaction push/pull
• ‘PDI’ - Paired Differences Index,
• ‘resource range’,
• ‘species specificity’,
• ‘PSI’ - pollination service index ,
• ‘betweenness’ - betweenness centrality and its weighted counterpart,
• ‘closeness’ - closeness centrality and its weighted counterpart,
• ‘Fisher’ - Fisher’s alpha index,
• ‘diversity’ - Shannon diversity of interactions
• ‘effective partners’,
• ‘proportional generality’,
• ‘proportional similarity’,
• ‘d’ - Blüthgen’s d’,
}
\examples{
\dontrun{
library(terra)
library(bipartite)
# load bipartite network and the raster stacks of higher level and lower level
species
bipnet <- read.csv(system.file("extdata", "bipnet.csv",
package="net.raster"), row.names=1, sep= ";" )#change separator if necessary
rasth <- rast(system.file("extdata", "rasth.tif",
package="net.raster"))
rastl <- rast(system.file("extdata", "rastl.tif",
package="net.raster"))
# applying the function to compute closeness centrality (default metric) for
lower level
lowclos <- specieslevel.spat (rasth, rastl, bipnet, level="lower")
plot(lowclos)
plot(lowclos$abuta_selloana) # you can choose one species at a time for
# better visualization
# calculating closeness, now for both levels (default)
allclos <- specieslevel.spat (rasth, rastl, bipnet)
plot(allclos$attila_rufus) # you can choose one species at a time
plot(allclos$abuta_selloana) #of each level
# computing normalised degree for higher level
hignd <- specieslevel.spat (rasth, rastl, bipnet, index="normalised degree",
level="higher")
plot(hignd$saltator_similis)# you can choose one species at a time
}
}
\references{
Carsten F. Dormann & Jochen Fründ ("bipartite" package)
Alarcon, R., Waser, N.M. and Ollerton, J. 2008. Year-to-year variation in the
topology of a plantpollinator interaction network. Oikos 117, 1796–1807
Barrat, A., M. Barthélemy, R. Pastor-Satorras, and A. Vespignani. 2004. The
architecture of complex weighted networks. Proceedings of the National Academy
of Sciences of the USA 101, 3747—3752. doi: 10.1073/pnas.0400087101.
Bascompte, J., Jordano, P. and Olesen, J. M. (2006) Asymmetric coevolutionary
networks facilitate biodiversity maintenance. Science 312, 431–433
Berlow, E. L., A. M. Neutel, J. E. Cohen, P. C. de Ruiter, B. Ebenman, M.
Emmerson, J. W. Fox,V. A. A. Jansen, J. I. Jones, G. D. Kokkoris, D. O.
Logofet, A. J. McKane, J. M. Montoya & O.
Petchey (2004) Interaction strengths in food webs: issues and opportunities.
Journal of Animal Ecology 73, 585-–598
Blüthgen, N., Menzel, F. and Blüthgen, N. (2006) Measuring specialization
in species interaction networks. BMC Ecology 6, 9
Dormann, C.F. (2011) How to be a specialist? Quantifying specialisation in
pollination networks.Network Biology 1, 1–20
Feinsinger, P., Spears, E.E. and Poole,R. W. (1981) A simple measure of
niche breadth. Ecology 62, 27–32.
Julliard, R., Clavel, J., Devictor, V., Jiguet, F. and Couvet, D. (2006)
Spatial segregation of specialists and generalists in bird communities.
Ecology Letters 9, 1237-–1244
Martín Gonzáles, A.M., Dalsgaard, B. and Olesen, J.M. (2010) Centrality
measures and the importance of generalist species in pollination networks.
Ecological Complexity, 7, 36–43
Opsahl, T. & Panzarasa, P. (2009). Clustering in weighted networks. Social
Networks, 31, 155–163
Poisot, T., Lepennetier, G., Martinez, E., Ramsayer, J., and Hochberg, M.E.
(2011a) Resource availability affects the structure of a natural bacteria-
bacteriophage community. Biology Letters 7,201–204
Poisot, T., Bever, J.D., Nemri, A., Thrall, P.H., and Hochberg, M.E. (2011b)
A conceptual framework for the evolution of ecological specialisation.
Ecology Letters 14, 841–851
Poisot, T., E. Canard, N. Mouquet, and M. E. Hochberg (2012) A comparative
study of ecological specialization estimators. Methods in Ecology and
Evolution 3, 537-–544. doi: 10.1111/j.2041-210X.2011.00174.x.
Vázquez, D. P., Melian, C. J., Williams, N. M., Blüthgen N., Krasnov B. R.
and Poulin, R. (2007) Species abundance and asymmetric interaction strength
in ecological networks. Oikos 116, 1120–1127
}
\seealso{
\code{\link{prep.web}}
}
\author{
Neander Marcel Heming and Cynthia Valéria Oliveira
}
