% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeModules_spat.R
\name{computMod_vec}
\alias{computMod_vec}
\title{ComputeModules_spat}
\usage{
computMod_vec(
  x,
  web,
  hlyr,
  method = "Beckett",
  deep = FALSE,
  deleteOriginalFiles = TRUE,
  steps = 1e+06,
  tolerance = 1e-10,
  experimental = FALSE,
  forceLPA = FALSE
)
}
\arguments{
\item{x}{A SpatRaster containing presence-absence data (0 or 1)
for a set of species. The layers (species) must be sorted according to the
tree order.}

\item{web}{A bipartite weighted graph}

\item{method}{Choice between the algorithm(s) provided by Stephen Beckett (2016) or Dormann & Strauss (2016) (\option{method="DormannStrauss"}). Defaults to the much faster and in the majority of cases better algorithm of Beckett. (Note the optional argument \option{forceLPA} to use his slightly inferior but even faster pure LPA algorithm.)}

\item{deep}{
If \code{deep} is set to FALSE (default), a flat clustering is computed, otherwise submodules are identified recursively within modules. Works only with \option{method="DormannnStrauss"}.
}

\item{deleteOriginalFiles}{
If \code{deleteOriginalFiles} is set to TRUE (default), the files mentioned above in the description are deleted from the hard drive disk, otherwise not. Applies only to \option{method="DormannnStrauss"}.
}

\item{steps}{
\code{steps} is the number of steps after which the computation of modules stops if no better division into modules than the current one can be found. Applies only to \option{method="DormannnStrauss"}.
}

\item{tolerance}{How small should the difference between MCMC-swap results be? At some point computer precision fluctuations make the algorithm fail to converge, which is why we choose a (very low) defaults of 1E-10. Applies only to \option{method="DormannnStrauss"}.}

\item{experimental}{Logical; using an undescribed and untested version for which no detail is available? (We suggest: not yet.) Applies only to \option{method="DormannnStrauss"}.}

\item{forceLPA}{Logical; should the even faster pure LPA-algorithm of Beckett be used? DIRT-LPA, the default, is less likely to get trapped in a local minimum, but is slightly slower. Defaults to FALSE.  Applies only to \option{method="Beckett"}.}
}
\value{
SpatRaster
}
\description{
Calculates modules by applying Newman's modularity measure to a
weighted bipartite network. The calculation will be made for each raster
cell, that is, for each subnetwork formed by the co-occurrence modeled for
the pairs of interacting species.
}
\references{
Rouven Strauss, with fixes by Carsten Dormann and Tobias Hegemann;
modified to accommodate Beckett’s algorithm by Carsten Dormann
}
\author{
Neander Marcel Heming and Cynthia Valéria Oliveira
}
