% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nested.spat.R
\name{nested.spat}
\alias{nested.spat}
\title{Several nestedness metrics for raster data}
\usage{
nested.spat(
  rh,
  rl,
  web,
  method = "weighted NODF",
  rescale = FALSE,
  normalised = TRUE
)
}
\arguments{
\item{rh}{SpatRaster. A raster (stack) containing presence-absence data (0 or 1)
for the higher level set of species.}

\item{rl}{A SpatRaster. A raster (stack) containing presence-absence data (0 or 1)
for the lower level set of species.}

\item{web}{Matrix. A weighted bipartite network matrix, binary (o or 1) or
not, where the lower level species (e.g. plants) are rows and higher level (e.g.
frugivores or pollinators)species are columns. The layers (species) of each
raster must be sorted according to the bipartite network order!}
}
\value{
Spatraster with the nestedness metric required
}
\description{
Calculates any of several nestedness metrics available for
bipartite networks, which can be weighted or not, for raster data. For a
time-efficient processing of spatial data, users may choose one of the nine
available nestedness indices ("discrepancy", "discrepancy2", "binmatnest",
"NODF", "NODF2", "C score", "checker","weighted NODF", "wine"). See more
about them at \link[bipartite]{nested}
}
\details{
Note that if a network is very small, with few nodes or links, it may be
impossible to calculate the choose metric or calculated result may be
unreliable.As the calculation is made with the subnet of each pixel, even in
cases like this, it is possible to visualize the spatialized metric on a
macroecological scale. See more about that and the available nestedness
indices at \link[bipartite]{nested}
}
\examples{
\dontrun{
library(terra)
library(bipartite)
# load bipartite network and the raster stacks of higher level and lower level
species
bipnet <- read.csv(system.file("extdata", "bipnet.csv",
package="net.raster"), row.names=1, sep= ";" )#change separator if necessary
rasth <- rast(system.file("extdata", "rasth.tif",
package="net.raster"))
rastl <- rast(system.file("extdata", "rastl.tif",
package="net.raster"))
# applying the function to compute weighted NODF (default)
wNODF <- nested.spat (rasth, rastl, bipnet)
plot(wNODF)
# applying the function to compute C score index
Cscore <- nested.spat (rasth, rastl, bipnet, method = "C score")
plot(Cscore)

}

}
\references{
Carsten F. Dormann ("bipartite" package)
Almeida-Neto, M., Guimaraes, P., Guimaraes, P.R., Loyola, R.D. and Ulrich,
W. 2008. A consistent metric for nestedness analysis in ecological systems:
reconciling concept and measurement. Oikos 117, 1227–1239.
Almeida-Neto, M. and Ulrich, W. (2011) A straightforward computational
approach for measuring nestedness using quantitative matrices. Environmental
Modelling & Software, 26, 173–178
Blüthgen, N., J. Fründ, D. P. Vazquez, and F. Menzel. 2008. What do
interaction network metrics tell us about specialisation and biological
traits? Ecology 89, 3387–3399.
Brualdi, R.A. and Sanderson, J.G. 1999. Nested species subsets, gaps, and
discrepancy. Oecologia 119, 256–264.
Felix, G.M., Pinheiro, R.B.P., Poulin, R., Krasnov, B.R. & Mello, M.A.R.
(2017). The compound topology of a continent-wide interaction network
explained by an integrative hypothesis of specialization. bioRxiv
Galeano, J., Pastor, J.M., Iriondo and J.M. 2008. Weighted-Interaction
Nestedness Estimator (WINE): A new estimator to calculate over frequency
matrices. arXiv 0808.3397v2 \link{physics.bio-ph}
Rodríguez-Gironés, M.A. and Santamaría, L. 2006. A new algorithm to calculate
the nestedness temperature of presence-absence matrices. J. Biogeogr. 33,
924–935.
Stone, L. and Roberts, A. 1990. The checkerboard score and species
distributions. Oecologia 85, 74–79.
Almeida-Neto, M. and Ulrich, W. 2010. A straightforward computational approach
for measuring nestedness using quantitative matrices. Environmental
Modelling & Software, in press.
}
\seealso{
\code{\link{prep.web}}, \code{\link{networklevel.spat}}
}
\author{
Neander Marcel Heming and Cynthia Valéria Oliveira
}
