% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/networklevel_spat.R
\name{networklevel.spat}
\alias{networklevel.spat}
\title{networklevel_spat metrics}
\usage{
networklevel.spat(
  rh,
  rl,
  web,
  index = "ALLBUTDD",
  level = "both",
  weighted = T,
  ISAmethod = "Bluethgen",
  SAmethod = "Bluethgen",
  extinctmethod = "r",
  nrep = 100,
  CCfun = "median",
  dist = "horn",
  normalise = T,
  empty.web = T,
  logbase = "e",
  intereven = "prod",
  H2_integer = T,
  fcweighted = T,
  fcdist = "euclidean",
  legacy = F
)
}
\arguments{
\item{web}{Web is a matrix representing the interactions observed between higher trophic level species (columns) and lower trophic level species (rows). 
Usually this will be number of pollinators on each species of plants or number of parasitoids on each species of prey.}

\item{index}{One or more of the following (exact match only!). First the group of \dQuote{pure} network indices, then those computed for each level.
   \itemize{
      \item \option{connectance},
      \item \option{web asymmetry},
      \item \option{links per species}, 
      \item \option{number of compartments}, 
      \item \option{compartment diversity}, 
      \item \option{cluster coefficient}, which will compute both the network-wide cluster coefficient as well as those for each level,
      \item \option{nestedness} (where ties may induce variation when changing the sequence of species in \option{web}), 
      \item \option{NODF},
      \item \option{weighted nestedness} (where ties may induce variation when changing the sequence of species in \option{web}), 
      \item \option{weighted NODF}, 
      \item \option{ISA} (or alternatively \option{interaction strength asymmetry} or 
      	\cr \option{dependence asymmetry}), 
      \item \option{SA} (or alternatively \option{specialisation asymmetry}),
      \item \option{linkage density}, 
      \item \option{weighted connectance},
      \item \option{Fisher alpha}, 
 %   \item \option{mean interaction diversity}, 
      \item \option{interaction evenness}, 
      \item \option{Alatalo interaction evenness}, 
      \item \option{Shannon diversity}, 
      \item \option{H2};  
   } and/or those invoked through \code{\link[bipartite]{grouplevel}}:
   \itemize{
     \item \option{number of species} in the respective trophic level,
     \item \option{mean number of links},
     \item \option{mean number of shared partners},  
     \item \option{weighted cluster coefficient},  
     \item \option{degree distribution}, 
     \item \option{togetherness}, 
     \item \option{C score}, 
     \item \option{V ratio}, 
     \item \option{discrepancy} (where ties may induce variation when changing the sequence of species in \option{web}), 
     \item \option{extinction slope} (where ties may induce variation when changing the sequence of species in \option{web}), 
     \item \option{robustness} (where ties may induce variation when changing the sequence of species in \option{web}),  
      \item \option{niche overlap},  
      \item \option{generality},
      \item \option{vulnerability},
      %\item \option{effective partners}, 
      \item \option{fc} (or alternatively \option{functional complementarity}).
    } Furthermore, there are some groups of indices that can be called:
    \itemize{
      \item \option{ALL} calculates all indices (returning a list (since degree distribution fits are a table within this list and cannot be simplified)), 
     \item \option{ALLBUTDD} (default) calculates all indices except degree distribution fits. This latter has the advantage that the output can be simplified to a vector; 
     \item \option{info} returns more general information on the network;
     \item \option{binary} returns a best-of selection of indices based on a binary network;
     \item \option{quantitative} returns a best-of selection of indices based on quantitative networks;
     \item \option{topology} returns indices more abstractly describing network properties. Also CHECK details below!
     }
   }

\item{level}{For which level should the level-specific indices be computed: \option{both} (default), \option{lower} or \option{higher}?}

\item{weighted}{Logical; should the weighted average be computed for indices that are averaged across species (at the group level)? Defaults to TRUE.}

\item{ISAmethod}{Method to use for calculating interaction strength (= dependence) asymmetry; original by \option{Bascompte} is yielding artefact results based only on the asymmetry of the web (as shown by example in Bl端thgen et al. 2007 analytically in Bl端thgen 2010) and should hence be avoided; \option{Bluethgen} (default) excludes singletons and corrects for low number of interactions (range -1 to 1).}

\item{SAmethod}{How to aggregate d'-based specialisation values: mean of log-transformed dependencies (\option{log}) or Bl端thgen's marginal totals-weighted mean (default); see Bl端thgen et al. (2007).}

\item{extinctmethod}{Specifies how species are removed from matrix: \option{random}, \option{degree} or \option{abundance} (partial matching). See \code{\link[bipartite]{second.extinct}} for details an option to predefine the sequence externally; idea from Memmott et al. (2004).}

\item{nrep}{Number of replicates for the extinction sequence analysis.}

\item{CCfun}{Method to use when calculating the clustering coefficient. Originally proposed as mean of cluster coefficients for each species. Defaults to \option{median}, because cluster coefficients are strongly skewed. }

\item{dist}{Distance metric to be used to calculate niche overlap. Any of \pkg{vegan}'s \code{vegdist}-metrics can be used; defaults to Horn's index, which is the recommendation of Krebs (1989). Binary percent niche overlap would be computed with \cr  \option{dist = "jaccard"}.}

\item{normalise}{Logical; shall the C-score and togetherness metrics be normalised to a range of 0 to 1? Defaults to \code{TRUE}.}

\item{empty.web}{Shall the empty columns and rows be deleted? Defaults to TRUE.}

\item{logbase}{Shall the various diversity indices (linkage density, partner diversity, generality/vulnerability, interaction evenness) be calculated to the base of e (default) or 2? Log2 is the proposal for generality and vulnerability by Bersier et al. (2002), while Shannon uses ln. The choice of the base will not affect the results qualitatively, at most by a scaling factor. Note that for all these indices, we follow common practice and define 0 * log(0) = 0.}

\item{intereven}{Shall all cells of the matrix be used to calculate the interaction evenness \cr
  (\option{intereven = "prod"})? Or, as given by Bersier et al. (2002) and Tylianakis et al. (2007), should only the realised links be used (\option{intereven = "sum"}; default)? Prod and sum refer to using the log of the \bold{product} of matrix dimensions (i.e. all cells) or the log of the \bold{sum} of non-zero cells (i.e. number of links)  as denominator in the evenness formula. See last paragraph of the details-section for views on these two options!}

\item{H2_integer}{Logical; indicates whether values in web are integers. Passed on to \code{\link[bipartite]{H2fun}}; see there for details.}

\item{fcweighted}{Logical; when computing ``functional complementarity'' sensu function \code{\link[bipartite]{fc}}, should the weights of the matrix be used. Defaults to TRUE, but original paper (Devoto et al. 2012) is based on FALSE.}

\item{fcdist}{Distance measure to be used to compute functional complementarity through \code{\link[bipartite]{fc}}; any measure accepted by \code{\link{dist}} is acceptable.}

\item{legacy}{Logical; should the old (pre-2.00) version of \code{networklevel} be used? To be backward compatible, the old \code{networklevel}-function is still available (\code{.networklevel}) and can be called by setting \option{legacy=TRUE}. This is only for the transition period until all papers in the making have been published (or binned). Index names and sometimes unclear focal level were downsides of the old implementation, which is now remedied. Thus, the use of \option{legacy=TRUE} and the direct call of \code{.networklevel} are strongly discouraged!}
}
\description{
rastrerize network level metrics
}
