% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.web.R
\name{prep.web}
\alias{prep.web}
\title{Prepares raster and network data}
\usage{
prep.web(rh, rl, web)
}
\arguments{
\item{rh}{SpatRaster. A raster (stack) containing presence-absence data (0 or 1)
for the higher level set of species.}

\item{rl}{A SpatRaster. A raster (stack) containing presence-absence data (0 or 1)
for the lower level set of species.}

\item{web}{Matrix. A weighted bipartite network matrix, binary (o or 1) or
not, where the lower level species (e.g. plants) are rows and higher level (e.g.
frugivores or pollinators)species are columns. The layers (species) of each
raster must be sorted according to the bipartite network order!}
}
\value{
A list with a matrix (subnetwork with the species present in rasters),
a logical vector indicating which species are from higher level among all,
and the two cleaned raster stacks of species.
}
\description{
Excludes in the network the species that not interact. Also checks
whether the species names of each Spatraster (higher and lower level) match
the species names in the cleaned network.
}
\examples{
\dontrun{
library(terra)
library(net.raster)
# load bipartite network and the raster stacks of higher level and lower level
species
bipnet <- read.csv(system.file("extdata", "bipnet.csv",
package="net.raster"), row.names=1)
rasth <- rast(system.file("extdata", "rasth.tif",
package="net.raster"))
rastl <- rast(system.file("extdata", "rastl.tif",
package="net.raster"))
# applying the function to check species names on all data
resu  <- prep.web (rasth, rastl, bipnet)
resu

}
}
\author{
Neander Marcel Heming and Cynthia Valéria Oliveira
}
